Keep MAX_BUILD_SPAN hours in memory, but cut away 'jobs' after DEFAULT_BUILD_SPAN
We will want to get the overall stats for the past 30 days, but not
necessarily the jobs data. By removing the memory footprint for the jobs
arrays, we can store the full 30 days in memory with minimal additional
memory footprint, as the jobs and individual job steps can account for
up to 85% of the total builds entry.